Add: button to go back to initial position after clicking a link
When click a link, the position is stored. In this case, display a
button to go back the position before click the link. When click this
button we poped out the position and go to the initial position.
This allow to navigate to a footnode and go back to the initial
position to continue reading.
The button is visible only when link stack is not empty.