Add: unstack if we click a backlink
When move to a footnote, sometimes a backlink yet exists.
If we click on it, we compare the destination to the last source
stored when click a link. If the last click was the link to the foot
note, the destination of the link and the id of the previous one can
be the same.
In this case, we move to the last position using the stack instead of
following the link.